+++ src/EpiIndicators.cpp
@@ -17,9 +17,31 @@
 #include "EpiIndicators.h"
 #include "utilities.h"
 
-#ifndef R_COMPILE
-#define R_COMPILE
-#endif // R_COMPILE
+//#ifndef R_COMPILE
+//#define R_COMPILE
+//#endif // R_COMPILE
+
+/// ------------------------------------------------------------------------------------------------------
+/// MEAN OF THE LAST ELEMENTS OF A VECTOR
+/// ------------------------------------------------------------------------------------------------------
+double mean_tail(
+    vector<double> &i /** input vector */,
+    int Ndays /** size of the tail to compute the mean */){
+  double mean=0;
+  for(int k=i.size()-Ndays;k<(int) i.size();k++) mean+=i[k];
+  return mean/Ndays;
+}
+
+/// ------------------------------------------------------------------------------------------------------
+/// GEOMETRIC MEAN OF THE LAST ELEMENTS OF A VECTOR
+/// ------------------------------------------------------------------------------------------------------
+double geometric_mean_tail(
+    vector<double> &i /** input vector */,
+    int Ndays /** size of the tail to compute the geometric mean */){
+  double mean=1;
+  for(int k=i.size()-Ndays;k<(int) i.size();k++) mean*=i[k];
+  return pow(mean,1./Ndays);
+}
 
 /// ------------------------------------------------------------------------------------------------------
 /// JOIN INDICATOR VALUES BY DATE
@@ -62,10 +84,8 @@ int joint_indicators_by_date(
         break;
       }
     }
-    //printf("current_date=%s %1.0lf %1.0lf \n",date[date.size()-1].c_str(),f[f.size()-1],g[g.size()-1]);
   }
   
-  //for(int k=0;k< (int) date1.size();k++) printf("%s\n",date1[k].c_str());
   
   return 0;
 }
@@ -284,6 +304,8 @@ double shift_and_mortality_estimation(
     double s_max /**  maximum value of the shift */,
     double wr /** regularity weight for the ratio */,
     double ws /** regularity weight for the shift */,
+    int tail /** tail of values used to fit f(t)*r(t)-g(t+s(t)) for the last values of the sequence */,
+    double tail_mu /** exponential coefficient to increase thw weight of the tail values */,
     double r_init /** optional value of the ratio on the left time interval extreme */,
     double r_end /** optional value of the ratio on the right time interval extreme */,
     double s_init /** optional value of the shift on the left time interval extreme */,
@@ -291,7 +313,7 @@ double shift_and_mortality_estimation(
 {
   int Nc=tmax-tmin+1;
   vector<double>  c1(c0.size()),d1(d0.size()),dp(d0.size());
-  vector<double>  c(Nc),d(Nc),s(Nc),r(Nc);
+  vector<double>  c(Nc),d(Nc),s(Nc),r(Nc),p(Nc,1.);
   
   double c0_median=percentil( c0.size()/2,c0);
   double d0_median=percentil( d0.size()/2,d0);
@@ -309,6 +331,12 @@ double shift_and_mortality_estimation(
   for(int k=0;k<Nc;k++){
     c[k]=c1[tmin+k];
     s[k]=s2[tmin+k];
+  }
+  
+  //int tail=56;
+  //double mu=0.1;
+  for(int k=Nc-tail;k<Nc;k++){
+    p[k]=exp((k-Nc+tail+1)*tail_mu);
   }
   
   double Error0=0,Error1=0,Error2=0,Error0N=0,Error1N=0,Error2N=0,Error=0,ErrorN=0,dif=1e20;
@@ -329,11 +357,11 @@ double shift_and_mortality_estimation(
     vector<double> diag(Nc,0.),sup(Nc,-wr),inf(Nc,-wr),b(Nc,0.);
     diag[0]=wr+c[0]*c[0];
     b[0]=c[0]*dn[0];
-    diag[Nc-1]=wr+c[Nc-1]*c[Nc-1];
-    b[Nc-1]=c[Nc-1]*dn[Nc-1];
+    diag[Nc-1]=wr+c[Nc-1]*c[Nc-1]*p[Nc-1];
+    b[Nc-1]=c[Nc-1]*dn[Nc-1]*p[Nc-1];
     for(int k=1;k<Nc-1;k++){
-      diag[k]=2*wr+c[k]*c[k];
-      b[k]=c[k]*dn[k];
+      diag[k]=2*wr+c[k]*c[k]*p[k];
+      b[k]=c[k]*dn[k]*p[k];
     }
     if(r_init!=-1e6){ diag[0]=1; sup[0]=0; b[0]=r_init/d0_median*c0_median; }
     if(r_end!=-1e6){ diag[Nc-1]=1; inf[Nc-2]=0; b[Nc-1]=r_end/d0_median*c0_median; }
@@ -347,7 +375,7 @@ double shift_and_mortality_estimation(
         double x=r[k]*c[k]-linear_interpolation2(d1,tmin+k+s[k]);
         double y=k==0?0:r[k]-r[k-1];
         double z=k==0?0:s[k]-s[k-1];
-        Error0+=x*x;
+        Error0+=p[k]*x*x;
         Error1+=wr*y*y;
         Error2+=ws*z*z;
       }
@@ -360,11 +388,11 @@ double shift_and_mortality_estimation(
     inf=vector<double>(Nc,-ws);
     diag[0]=ws+dpn[0]*dpn[0];
     b[0]=(r[0]*c[0]-dn[0])*dpn[0]-ws*(s[0]-s[1]);
-    diag[Nc-1]=ws+dpn[Nc-1]*dpn[Nc-1];
-    b[Nc-1]=(r[Nc-1]*c[Nc-1]-dn[Nc-1])*dpn[Nc-1]-ws*(s[Nc-1]-s[Nc-2]);
+    diag[Nc-1]=ws+dpn[Nc-1]*dpn[Nc-1]*p[Nc-1];
+    b[Nc-1]=(r[Nc-1]*c[Nc-1]-dn[Nc-1])*dpn[Nc-1]*p[Nc-1]-ws*(s[Nc-1]-s[Nc-2]);
     for(int k=1;k<Nc-1;k++){
-      diag[k]=2*ws+dpn[k]*dpn[k];
-      b[k]=(r[k]*c[k]-dn[k])*dpn[k]-ws*(2*s[k]-s[k-1]-s[k+1]);
+      diag[k]=2*ws+dpn[k]*dpn[k]*p[k];
+      b[k]=(r[k]*c[k]-dn[k])*dpn[k]*p[k]-ws*(2*s[k]-s[k-1]-s[k+1]);
     }
     
     for(int k=0;k<Nc;k++){
@@ -392,7 +420,7 @@ double shift_and_mortality_estimation(
       double x=r[k]*c[k]-linear_interpolation2(d1,tmin+k+s[k]);
       double y=k==0?0:r[k]-r[k-1];
       double z=k==0?0:s[k]-s[k-1];
-      Error0N+=x*x;
+      Error0N+=p[k]*x*x;
       Error1N+=wr*y*y;
       Error2N+=ws*z*z;
       if(s[k]>s_max) Error3N+=1e10*(s_max-s[k])*(s_max-s[k]);
@@ -433,49 +461,88 @@ double shift_and_mortality_estimation(
 }
 
 /// ------------------------------------------------------------------------------------------------------
+/// FORECASTING THE INDICATOR f(t) FROM THE INDICATOR g(t)
+/// ------------------------------------------------------------------------------------------------------
+void indicators_forecast(
+    vector<string> &dates /** date for the first indicator value (input/output) */,
+    vector<double> &f /** first epidemiological indicator (input/output) */,
+    vector<double> &g /** second epidemiological indicator with forecast (input) */,
+    vector<double> &r0 /** ratio between indicators (input) */,
+    vector<double> &s0 /** shift between indicators (input) */,
+    double r_forecast_end /** final value of r(t) for the Hermite interpolation (input)*/,
+    double s_forecast_end /** final value of s(t) for the Hermite interpolation (input) */,
+    int NforecastDays /** number of forecast days for f(t) (input) */,
+    int NdaysHermite /** Number of days used for the Hermite main component (input) */,
+    int NdaysHermiteDerivada /** Number of days used for the Hermite derivative component (input) */
+)
+{
+  vector<double> r=r0,s=s0;
+  
+  HermiteInterpolation(r,r_forecast_end,NdaysHermite,NdaysHermiteDerivada);
+  HermiteInterpolation(s,s_forecast_end,NdaysHermite,NdaysHermiteDerivada);
+  int Ndays=f.size()+NforecastDays;
+  for(int k=f.size();k<Ndays;k++){
+    double si=linear_interpolation2(s,(double) k);
+    double ri=linear_interpolation2(r,(double) k);
+    double gi=linear_interpolation2(g,(double) k+si);
+    f.push_back(gi/ri);
+    time_t temp=string2date(dates[dates.size()-1].c_str());
+    temp+=86400+1000;
+    dates.push_back(date2string(temp));
+  }
+  
+  r0=r;
+  s0=s;
+}
+
+
+/// ------------------------------------------------------------------------------------------------------
 /// OPTIMIZATION OF THE RATIO AND SHIFT BETWEEN INDICATORS USING DIFFERENT INITIALIZATIONS OF s(t)
 /// ------------------------------------------------------------------------------------------------------
 double shift_and_ratio_optimization(
     vector<string> &dates /** date for each indicator value */,
-    vector<double> &c /** first epidemiological indicator sequence */,
-    vector<double> &d /** second epidemiological indicator sequence */,
+    vector<double> &f /** first epidemiological indicator sequence */,
+    vector<double> &g /** second epidemiological indicator sequence */,
     vector<double> &r /** output ratio between indicators */,
     vector<double> &s /** output shift between indicators */,
     int s_min /** minimum value of the shift */,
     int s_max /** maximum value of the shift */,
     double wr /** regularity weight for the ratio */,
     double ws /** regularity weight for the shift */,
+    int tail /** tail of the dates used to fit f(t)*r(t)-g(t+s(t)) for the last values of the sequence */,
+    double tail_mu /** exponential coefficient to increase the weight of the tail values */,
     double r_init /** optional value of the ratio on the left time interval extreme */,
     double r_end /** optional value of the ratio on the right time interval extreme */,
     double s_init /** optional value of the shift on the left time interval extreme */,
     double s_end /** optional value of the shift on the right time interval extreme */)
 {
-  while( (-s_min>0 && -s_min< (int) c.size() && c[(int) -s_min]==0)
-           || (s_max>0 && s_max< (int) d.size() && d[s_max]==0) ){
+  while( (-s_min>0 && -s_min< (int) f.size() && f[(int) -s_min]==0)
+           || (s_max>0 && s_max< (int) g.size() && g[s_max]==0) ){
     dates.erase(dates.begin());
-    c.erase(c.begin());
-    d.erase(d.begin());
+    f.erase(f.begin());
+    g.erase(g.begin());
   }
-  int Nc=c.size()-1;
-  while( (Nc-s_max>0 && Nc-s_max<  (int) c.size() && c[Nc-s_max]==0) ||
-         (Nc+s_min>0 && Nc+s_min< (int) d.size() && d[Nc+s_min]==0) || d[Nc]==0){
+  //printf("tail=%d tail_mu=%lf\n",tail,tail_mu);
+  int Nc=f.size()-1;
+  while( (Nc-s_max>0 && Nc-s_max<  (int) f.size() && f[Nc-s_max]==0) ||
+         (Nc+s_min>0 && Nc+s_min< (int) g.size() && g[Nc+s_min]==0) || g[Nc]==0){
     dates.resize(Nc);
-    c.resize(Nc);
-    d.resize(Nc--);
+    f.resize(Nc);
+    g.resize(Nc--);
   }
   
-  if(c.size()<56){
+  if(f.size()<56){
     return -1.;
   }
   
-  Nc=c.size()/56;
+  Nc=f.size()/56;
   double min_error=2e20;
   double max_ratio_left=0.1;
   double max_ratio_right=0.1;
   int radius=28;
   for(int type=0;type<=Nc;type++){
-    vector<double> r2,s2=shift_initialization(c,d,radius,max_ratio_left,max_ratio_right,s_min,s_max,type);
-    double error=shift_and_mortality_estimation(c,d,r2,s2,0,c.size()-1,s_min,s_max,wr,ws,r_init,r_end,s_init,s_end);
+    vector<double> r2,s2=shift_initialization(f,g,radius,max_ratio_left,max_ratio_right,s_min,s_max,type);
+    double error=shift_and_mortality_estimation(f,g,r2,s2,0,f.size()-1,s_min,s_max,wr,ws,tail,tail_mu,r_init,r_end,s_init,s_end);
     if(error>=0 && error<min_error){
       min_error=error;
       r=r2;